fix(create-zudo-doc): provide diff peer unconditionally so barebone projects build
packageOwnedRoutes defaults ON since 1.0, so the always-copied host base
template pages/lib/_doc-history-area.tsx statically imports the real DocHistory
from @takazudo/zudo-doc/doc-history (island-scanner chain page→stub→DocHistory),
pulling dist/doc-history/index.js's `await import("diff")` into EVERY generated
bundle — even a docHistory-off / barebone project. The scaffold only added the
`diff` peer dep when docHistory was enabled, so a barebone `zfb build` failed at
esbuild with `Could not resolve "diff"` (the live tail of the #2342 chain).

Move `diff` to the unconditional generated deps, next to the other always-bundled
peers (preact/shiki/zod/katex). Audit confirmed `diff` is the only mis-gated
always-bundled peer; history-server and zdtp stay correctly feature-gated.

Add a true-barebone (all features off) scaffold→install→`zfb build` slow test
that locks the regression — the previously-unprotected case (preset-swap only
covers i18n-on). Extract the shared install-retry/build plumbing into
slow-build-helpers.ts, reused by both slow tests.

Refs #2342. Release intentionally not bumped (lands on main unreleased).
